refactor: divide types and schemes by subject
