+++ apps/server/src/api/auth.ts
@@ -1,5 +1,5 @@
 import { Router, type NextFunction, type Request, type Response } from 'express';
-import { LoginDtoSchema, RegisterDtoSchema } from '@repo/shared/src/schemas.ts';
+import { LoginDtoSchema, RegisterDtoSchema } from '@repo/shared/src/schemas/auth.ts';
 import { Endpoints, HttpStatus, ServerConstants } from '@repo/shared/src/api.constants.ts';
 import { prisma } from '../prisma/prisma.ts';
 import * as argon from 'argon2';
