+++ packages/shared/src/schemas/user.ts
@@ -1,0 +1,8 @@
+import z from 'zod';
+
+export const UserSchema = z.object({
+  id: z.string(),
+  email: z.string(),
+  username: z.string(),
+  password: z.string(),
+});
