+++ packages/shared/src/types/auth.ts
@@ -1,15 +1,8 @@
-import z from 'zod';
-import type {
-  LoginDtoSchema,
-  RegisterDtoSchema,
-  UserSchema,
-  WSHandshakeAuthSchema,
-} from './schemas';
+import * as z from 'zod';
+import type { LoginDtoSchema, RegisterDtoSchema, WSHandshakeAuthSchema } from '../schemas/auth';
 
 export type LoginDto = z.infer<typeof LoginDtoSchema>;
 
 export type RegisterDto = z.infer<typeof RegisterDtoSchema>;
-
-export type User = z.infer<typeof UserSchema>;
 
 export type WSHandshakeAuth = z.infer<typeof WSHandshakeAuthSchema>;
