+++ packages/shared/src/types/user.ts
@@ -1,0 +1,4 @@
+import * as z from 'zod';
+import type { UserSchema } from '../schemas/user';
+
+export type User = z.infer<typeof UserSchema>;
